Fix overlap = 0 in some cases
I was running into some bugs with response.overlap being equal to 0 in some cases which also messes up overlap_v. After running some tests, it turns out that overlap can sometimes be equal to 0 which is due to the fact we're testing against all normals. Which causes a problem if you're replacing response.overlap with the lowest value. Checking if abs_overlap is greater than 0 (line 202) fixed this for me completely.